flask/requirements.txt
appnope==0.1.0
backcall==0.2.0
BaseModel==20200715.1219
click==7.1.2
decorator==4.4.2
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.8
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jupyter-client==6.1.6
jupyter-core==4.6.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pydantic==1.6
Pygments==2.6.1
pymongo==3.11.0
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.14.0
pyzmq==19.0.2
six==1.15.0
SQLAlchemy==1.3.18
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.5
Werkzeug==1.0.1
